refactor: split unit tests by GitLab API resources
