+++ gitlab/tests/objects/test_groups.py
@@ -1,4 +1,8 @@
-import unittest
+"""
+GitLab API: https://docs.gitlab.com/ce/api/groups.html
+"""
+
+import pytest
 
 from httmock import response, urlmatch, with_httmock
 
@@ -36,66 +40,55 @@ def resp_create_import(url, request):
     return response(202, content, headers, None, 25, request)
 
 
-class TestGroup(unittest.TestCase):
-    def setUp(self):
-        self.gl = gitlab.Gitlab(
-            "http://localhost",
-            private_token="private_token",
-            ssl_verify=True,
-            api_version=4,
-        )
-
-    @with_httmock(resp_get_group)
-    def test_get_group(self):
-        data = self.gl.groups.get(1)
-        assert isinstance(data, gitlab.v4.objects.Group)
-        assert data.name == "name"
-        assert data.path == "path"
-        assert data.id == 1
-
-    @with_httmock(resp_create_group)
-    def test_create_group(self):
-        name, path = "name", "path"
-        data = self.gl.groups.create({"name": name, "path": path})
-        assert isinstance(data, gitlab.v4.objects.Group)
-        assert data.name == name
-        assert data.path == path
+@with_httmock(resp_get_group)
+def test_get_group(gl):
+    data = gl.groups.get(1)
+    assert isinstance(data, gitlab.v4.objects.Group)
+    assert data.name == "name"
+    assert data.path == "path"
+    assert data.id == 1
 
 
-class TestGroupExport(TestGroup):
-    def setUp(self):
-        super(TestGroupExport, self).setUp()
-        self.group = self.gl.groups.get(1, lazy=True)
-
-    @with_httmock(resp_create_export)
-    def test_create_group_export(self):
-        export = self.group.exports.create()
-        assert export.message == "202 Accepted"
-
-    @unittest.skip("GitLab API endpoint not implemented")
-    @with_httmock(resp_create_export)
-    def test_refresh_group_export_status(self):
-        export = self.group.exports.create()
-        export.refresh()
-        assert export.export_status == "finished"
-
-    @with_httmock(resp_create_export, resp_download_export)
-    def test_download_group_export(self):
-        export = self.group.exports.create()
-        download = export.download()
-        assert isinstance(download, bytes)
-        assert download == binary_content
+@with_httmock(resp_create_group)
+def test_create_group(gl):
+    name, path = "name", "path"
+    data = gl.groups.create({"name": name, "path": path})
+    assert isinstance(data, gitlab.v4.objects.Group)
+    assert data.name == name
+    assert data.path == path
 
 
-class TestGroupImport(TestGroup):
-    @with_httmock(resp_create_import)
-    def test_import_group(self):
-        group_import = self.gl.groups.import_group("file", "api-group", "API Group")
-        assert group_import["message"] == "202 Accepted"
+@with_httmock(resp_create_export)
+def test_create_group_export(group):
+    export = group.exports.create()
+    assert export.message == "202 Accepted"
 
-    @unittest.skip("GitLab API endpoint not implemented")
-    @with_httmock(resp_create_import)
-    def test_refresh_group_import_status(self):
-        group_import = self.group.imports.get()
-        group_import.refresh()
-        assert group_import.import_status == "finished"
+
+@pytest.mark.skip("GitLab API endpoint not implemented")
+@with_httmock(resp_create_export)
+def test_refresh_group_export_status(group):
+    export = group.exports.create()
+    export.refresh()
+    assert export.export_status == "finished"
+
+
+@with_httmock(resp_create_export, resp_download_export)
+def test_download_group_export(group):
+    export = group.exports.create()
+    download = export.download()
+    assert isinstance(download, bytes)
+    assert download == binary_content
+
+
+@with_httmock(resp_create_import)
+def test_import_group(gl):
+    group_import = gl.groups.import_group("file", "api-group", "API Group")
+    assert group_import["message"] == "202 Accepted"
+
+
+@pytest.mark.skip("GitLab API endpoint not implemented")
+@with_httmock(resp_create_import)
+def test_refresh_group_import_status(group):
+    group_import = group.imports.get()
+    group_import.refresh()
+    assert group_import.import_status == "finished"
